♻️ Refactor: 테마별 영화 목록을 보여주는 코드를 MovieList 컴포넌트로 분리
